+++ shaderpack-superduper/shaders/main/deferred1.glsl
@@ -211,7 +211,9 @@
     #include "/lib/utility/noiseFunctions.glsl"
 
     #include "/lib/atmospherics/fogRender.glsl"
-    
+
+    #include "/lib/atmospherics/skyRender.glsl"
+
     #include "/lib/rayTracing/rayTracer.glsl"
 
     #include "/lib/lighting/complexShadingDeferred.glsl"